requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.26.129
botocore==1.29.129
celery==5.2.7
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==4.2.1
django-ses==3.4.1
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.31
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
s3transfer==0.6.1
six==1.16.0
slack-sdk==3.21.3
sqlparse==0.4.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
